Move shared code into separate project
- Instead of embedding shared code, make a new project
- Update namespaces accordingly
- Start abstracting away graphics separately from the game
