+++ StardewModding/StardewModding.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HarderBetterFasterStrongerA
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Predictor", "Predictor\Predictor.csproj", "{3EC62EED-CAD6-407A-9D99-CA18E767E43B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utility", "Utility\Utility.csproj", "{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{3EC62EED-CAD6-407A-9D99-CA18E767E43B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3EC62EED-CAD6-407A-9D99-CA18E767E43B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3EC62EED-CAD6-407A-9D99-CA18E767E43B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
