+++ StardewModding/Utility/Abstraction/StardewManagedPropertyAttribute.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 
-namespace Dannnno.StardewMods.Predictor.Shared
+namespace Dannnno.StardewMods.Abstraction
 {
     /// <summary>
     /// Attribute that indicates a property is managed in the StardewValley game
     /// </summary>
     [AttributeUsage(AttributeTargets.Property)]
-    public class StardewManagedPropertyAttribute : Attribute { }
+    internal class StardewManagedPropertyAttribute : Attribute { }
 }
