+++ StardewModding/Utility/UI/IStardewGraphics.cs
@@ -1,0 +1,10 @@
+﻿using Dannnno.StardewMods.Abstraction;
+
+namespace Dannnno.StardewMods.UI.Shared
+{
+    public interface IStardewGraphics
+    {
+
+        public IStardewGame Game { get; }
+    }
+}
