+++ StardewModding/Utility/Utility.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{3EC62EED-CAD6-407A-9D99-CA18E767E43B}</ProjectGuid>
+    <ProjectGuid>{6A49CAC1-0180-4319-BF9A-6A3FFBDCC142}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Dannnno.StardewMods.Predictor</RootNamespace>
-    <AssemblyName>Predictor</AssemblyName>
+    <RootNamespace>Dannnno.StardewMods</RootNamespace>
+    <AssemblyName>Utility</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -38,9 +38,6 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.ValueTuple, Version=4.0.3.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
-      <HintPath>..\packages\System.ValueTuple.4.5.0\lib\netstandard1.0\System.ValueTuple.dll</HintPath>
-    </Reference>
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -49,31 +46,27 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Geodes\GeodePredictor.cs" />
-    <Compile Include="Geodes\IGeodeTreasureCalculator.cs" />
-    <Compile Include="Geodes\StardewGeodeCalculator.cs" />
-    <Compile Include="UI\Geode\GeodeResultModel.cs" />
-    <Compile Include="UI\PredictorMenuV2.cs" />
-    <Compile Include="UI\Geode\GeodeMenuItem.cs" />
-    <Compile Include="UI\Geode\GeodeMenuTab.cs" />
-    <Compile Include="UI\PredictorMenu.cs" />
-    <Compile Include="Geodes\IGeodeService.cs" />
-    <Compile Include="Geodes\StardewGeodeService.cs" />
-    <Compile Include="ModEntry.cs" />
+    <Compile Include="Abstraction\GameStateContextManager.cs" />
+    <Compile Include="Abstraction\IStardewGame.cs" />
+    <Compile Include="Abstraction\IStardewObjectProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Abstraction\StardewDataObjectInfoProvider.cs" />
+    <Compile Include="Abstraction\StardewGameWrapper.cs" />
+    <Compile Include="Abstraction\StardewManagedPropertyAttribute.cs" />
+    <Compile Include="UI\ClickableMenu.cs" />
+    <Compile Include="UI\HoverTextBox.cs" />
+    <Compile Include="UI\IStardewGraphics.cs" />
+    <Compile Include="UI\MenuTab.cs" />
+    <Compile Include="UI\MenuTabItem.cs" />
+    <Compile Include="UI\MenuVerticalScrollBar.cs" />
+    <Compile Include="UI\TabbedClickableMenu.cs" />
   </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
-    <None Include="manifest.json" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\Pathoschild.Stardew.ModBuildConfig.2.2.0\analyzers\dotnet\cs\StardewModdingAPI.ModBuildConfig.Analyzer.dll" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Utility\Utility.csproj">
-      <Project>{6a49cac1-0180-4319-bf9a-6a3ffbdcc142}</Project>
-      <Name>Utility</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Pathoschild.Stardew.ModBuildConfig.2.2.0\build\Pathoschild.Stardew.ModBuildConfig.targets" Condition="Exists('..\packages\Pathoschild.Stardew.ModBuildConfig.2.2.0\build\Pathoschild.Stardew.ModBuildConfig.targets')" />
